Isolate operator utils
